examples: deduplicate examples
